# backend/src/main/java/com/example/sessionauth/dto/PersonDTO.java
package com.example.sessionauth.dto;

import lombok.Data;

@Data
public class PersonDTO {
    private String name;
    private String surname;
    private String patronymic;
    private String phone;
}