# backend/src/main/resources/swagger.yaml
openapi: 3.0.4
info:
  title: Session Auth API - Document Management
  description: API for managing documents in the Session Authentication system
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /documents/create:
    post:
      tags:
        - Document
      summary: Create a new document
      description: Creates a new document with the provided details. Requires EMPLOYEE authority.
      operationId: createDocument
      security:
        - bearerAuth: [EMPLOYEE]
      requestBody:
        description: Document to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentDTO'
      responses:
        '201':
          description: Document created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /documents/{id}:
    get:
      tags:
        - Document
      summary: Get document by ID
      description: Returns a single document by its ID. Requires EMPLOYEE authority.
      operationId: getDocumentById
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of document to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: Document not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Document
      summary: Update an existing document
      description: Updates an existing document. Requires EMPLOYEE authority.
      operationId: updateDocument
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of document to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Document data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentDTO'
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '400':
          description: Invalid ID or input
        '404':
          description: Document not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Document
      summary: Delete a document
      description: Deletes a document by ID. Requires EMPLOYEE authority.
      operationId: deleteDocument
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of document to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Document deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Document not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /documents/getAll:
    get:
      tags:
        - Document
      summary: Get all documents
      description: Returns a list of all documents. Requires EMPLOYEE authority.
      operationId: getAllDocuments
      security:
        - bearerAuth: [EMPLOYEE]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/create:
    post:
      tags:
        - Employee
      summary: Create a new employee
      description: Creates a new employee with the provided details. Requires ADMIN authority.
      operationId: createEmployee
      security:
        - bearerAuth: [ADMIN]
      requestBody:
        description: Employee to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDTO'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/{id}:
    get:
      tags:
        - Employee
      summary: Get employee by ID
      description: Returns a single employee by its ID. Requires ADMIN authority.
      operationId: getEmployeeById
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of employee to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: Employee not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Employee
      summary: Update an existing employee
      description: Updates an existing employee. Requires ADMIN authority.
      operationId: updateEmployee
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of employee to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Employee data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDTO'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDTO'
        '400':
          description: Invalid ID or input
        '404':
          description: Employee not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Employee
      summary: Delete an employee
      description: Deletes an employee by ID. Requires ADMIN authority.
      operationId: deleteEmployee
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of employee to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Employee deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Employee not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/getAll:
    get:
      tags:
        - Employee
      summary: Get all employees
      description: Returns a list of all employees. Requires ADMIN authority.
      operationId: getAllEmployees
      security:
        - bearerAuth: [ADMIN]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /institutes/create:
    post:
      tags:
        - Institute
      summary: Create a new institute
      description: Creates a new institute with the provided details. Requires EMPLOYEE authority.
      operationId: createInstitute
      security:
        - bearerAuth: [EMPLOYEE]
      requestBody:
        description: Institute to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstituteDTO'
      responses:
        '201':
          description: Institute created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstituteDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /institutes/{id}:
    get:
      tags:
        - Institute
      summary: Get institute by ID
      description: Returns a single institute by its ID. Requires EMPLOYEE authority.
      operationId: getInstituteById
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of institute to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstituteDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: Institute not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Institute
      summary: Update an existing institute
      description: Updates an existing institute. Requires EMPLOYEE authority.
      operationId: updateInstitute
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of institute to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Institute data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstituteDTO'
      responses:
        '200':
          description: Institute updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstituteDTO'
        '400':
          description: Invalid ID or input
        '404':
          description: Institute not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Institute
      summary: Delete an institute
      description: Deletes an institute by ID. Requires EMPLOYEE authority.
      operationId: deleteInstitute
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of institute to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Institute deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Institute not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /institutes/getAll:
    get:
      tags:
        - Institute
      summary: Get all institutes
      description: Returns a list of all institutes. Requires EMPLOYEE authority.
      operationId: getAllInstitutes
      security:
        - bearerAuth: [EMPLOYEE]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstituteDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    DocumentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          readOnly: true
        name:
          type: string
          example: "document.pdf"
          description: Name of the document
        bytes:
          type: string
          format: byte
          example: "base64-encoded-content"
          description: Document content as base64 encoded string
        userId:
          type: integer
          format: int64
          example: 123
          description: ID of the user who owns the document
      required:
        - name
        - bytes
        - userId
    EmployeeDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          readOnly: true
          description: Employee ID
        person:
          $ref: '#/components/schemas/PersonDTO'
        jobTitleId:
          type: integer
          format: int64
          example: 2
          description: ID of the job title
        userId:
          type: integer
          format: int64
          example: 3
          description: ID of the associated user
...
    PersonDTO:
      type: object
      properties:
        name:
          type: string
          example: "John"
          description: First name
        surname:
          type: string
          example: "Doe"
          description: Last name
        patronymic:
          type: string
          example: "Smith"
          description: Middle name (optional)
          nullable: true
        phone:
          type: string
          example: "+1234567890"
          description: Phone number
    InstituteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          readOnly: true
          description: Institute ID
        name:
          type: string
          example: "Tech Institute"
          description: Name of the institute
        email:
          type: string
          format: email
          example: "contact@techinstitute.com"
          description: Contact email
        phone:
          type: string
          example: "+1234567890"
          description: Contact phone number
      required:
        - name
        - email
        - phone
...
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input data"
        path:
          type: string
          example: "/api/v1/documents"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication